Don't skip tasks tagged [always,b] without '--skip-tags always'
The documentation (playbooks_tags.rst) suggests that tasks tagged with
'always' will be skipped only if you explicitly specify 'always' as a
tag to skip, but the code behaved differently. This problem becomes
especially noticeable when using the new "apply" feature to apply a
single tag across include_{tasks,roles}, because there's no way to
permit always-tagged included tasks to run.

As a side-effect of how the code was written, ``--skip-tags all,b``
would cause the task tagged with [always,b] to NOT be skipped. This is
unintuitive and confusing, so it's not an acceptable workaround for the
problem.

Reported upstream as https://github.com/ansible/ansible/issues/45045

Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>