Add a wait_for_ssh module
This module uses the normal ssh transport, so it respects ssh options
defined in the inventory (especially relevant to ProxyCommand), which
``wait_for`` running on localhost will not do.

This is modelled along the lines of wait_for_connection, but does not
depend on Python being installed on the target (because it uses raw to
execute a command instead of executing the ping module).

In future, this could be implemented as a ``test: raw`` parameter to
the wait_for_connection module upstream.

Signed-off-by: Abhijit Menon-Sen <ams@2ndQuadrant.com>