Change the wait_for_ssh test to be more permissive of extra text
Requiring the command to output 'pong' and nothing else (i.e., making no
allowances for /etc/issue.net) is overly restrictive, and could lead to
raw_test failures and timeouts on instances that are reachable. All we
really need is to ensure that the 'pong' is somewhere in the output.

At the same time, we switch to a timestamped 'pong NNN' string too. The
output string is still required to be on one line, all by itself (and on
stdout).