Compare SCRAM-SHA-256 encrypted passwords directly with rolpassword
When trying to decide if a password change is needed, passwords starting
with SCRAM-SHA-256$ can be compared directly to rolpassword.

The upstream postgresql_user module in community.general now has SCRAM
support, but it accepts *unencrypted* passwords, and constructs a SCRAM
verifier based on rolpassword for a comparison, but we don't need that
because we're always going to be passing an already-encrypted password
to this module.